requirements.txt
absl-py==0.1.5
enum34==1.1.6
future==0.16.0
glfw==1.4.0
lxml==4.1.1
mock==2.0.0
nose==1.3.7
numpy==1.13.3
pyparsing==2.2.0
six==1.11.0
